Aggregate a subset of results

Aggregate a subset of results

aggregate_results_subset(misty.results, folders)

Arguments

misty.results
a results list generated by collect_results().

folders
Paths to folders containing the raw results from run_misty().

Value

the misty.results list with an added list item importances.aggregated.subset containing the aggregated importances for the subset of folders.

See also