Collect and aggregate results

Collect and aggregate performance, contribution and importance estimations of a set of raw results produced by run_misty().

collect_results(folders)

Arguments

folders
Paths to folders containing the raw results from run_misty().

Value

List of collected performance, contributions and importances per sample, performance and contribution statistics and aggregated importances.

improvements

Long format tibble with measurements of performance for each target and each sample. Available performance measures are RMSE and variance explained (R2) for a model containing only an intrinsic view (intra.RMSE, intra.R2), model with all views (multi.RMSE, multi.R2), gain of RMSE and gain of variance explained of multi-view model over the intrisic model where gain.RMSE is the relative decrease of RMSE in percent, while gain.R2 is the absolute increase of variance explained in percent. Each value represents the mean performance across folds (k-fold cross-validation). The p values of a one sided t-test of improvement of performance (p.RMSE, p.R2) are also available as a measure.

improvements.stats

Long format tibble with summary statistics (mean, standard deviation and coefficient of variation) for all performance measures for each target over all samples.

contributions

Long format tibble with the values of the coefficients for each view in the meta-model, for each target and each sample. The p values for the coefficient for each view, under the null hypothesis of zero contribution to the meta model are also available.

contributions.stats

Long format tibble with summary statistics for all views per target over all samples. Including mean coffecient value, fraction of contribution, mean and standard deviation of p values.

importances

List of view-specific predictor-target importance tables per sample. The importances in each table are standardized per target and weighted by the quantile of the coefficient for the target in that view. Columns other than Predictor represent target markers.

importances.aggregated

A list of aggregated view-specific predictor-target importance tables . Aggregation is reducing by mean over all samples.

See also

Examples

# Train and collect results for 3 samples in synthetic

library(dplyr)
library(purrr)

misty.results <- synthetic[seq_len(3)] %>% 
                   imap_chr(~ create_initial_view(.x %>% select(-c(row,col,type))) %>%
                   add_paraview(.x %>% select(row,col), l = 10) %>% 
                   run_misty(paste0("results/", .y))) %>%
                   collect_results()

#> Generating paraview

#> Training models

#> Generating paraview

#> Training models

#> Generating paraview

#> Training models

#> Collecting results from results/synthetic1 results/synthetic10 results/synthetic2

#> 
#> Collecting improvements

#> 
#> Collecting contributions

#> 
#> Collecting importances

#> 
#> Aggregating

str(misty.results)

#> List of 6
#>  $ improvements          : tibble[,4] [264 × 4] (S3: tbl_df/tbl/data.frame)
#>   ..$ target : chr [1:264] "ECM" "ECM" "ECM" "ECM" ...
#>   ..$ sample : chr [1:264] "results/synthetic1" "results/synthetic1" "results/synthetic1" "results/synthetic1" ...
#>   ..$ measure: chr [1:264] "intra.RMSE" "intra.R2" "multi.RMSE" "multi.R2" ...
#>   ..$ value  : num [1:264] 0.10442 0.92634 0.09783 0.93532 0.00968 ...
#>  $ improvements.stats    : tibble[,5] [66 × 5] (S3: tbl_df/tbl/data.frame)
#>   ..$ target : chr [1:66] "ECM" "ECM" "ECM" "ECM" ...
#>   ..$ measure: chr [1:66] "gain.R2" "gain.RMSE" "intra.R2" "intra.RMSE" ...
#>   ..$ mean   : num [1:66] 0.857 6.202 0.929 0.1 0.937 ...
#>   ..$ sd     : num [1:66] 0.06568 0.66656 0.00335 0.00541 0.00366 ...
#>   ..$ cv     : num [1:66] 0.0766 0.10747 0.00361 0.05395 0.00391 ...
#>  $ contributions         : tibble[,4] [198 × 4] (S3: tbl_df/tbl/data.frame)
#>   ..$ target: chr [1:198] "ECM" "ECM" "ECM" "ECM" ...
#>   ..$ sample: chr [1:198] "results/synthetic1" "results/synthetic1" "results/synthetic1" "results/synthetic1" ...
#>   ..$ view  : chr [1:198] "intercept" "intra" "para.10" "p.intercept" ...
#>   ..$ value : num [1:198] -7.00e-02 9.85e-01 2.49e-01 6.93e-98 0.00 ...
#>  $ contributions.stats   : tibble[,6] [22 × 6] (S3: tbl_df/tbl/data.frame)
#>   ..$ target  : chr [1:22] "ECM" "ECM" "ligA" "ligA" ...
#>   ..$ view    : chr [1:22] "intra" "para.10" "intra" "para.10" ...
#>   ..$ mean    : num [1:22] 0.9887 0.2449 0.9965 0.0672 0.9858 ...
#>   ..$ fraction: num [1:22] 0.8015 0.1985 0.9368 0.0632 0.7744 ...
#>   ..$ p.mean  : num [1:22] 0.00 3.91e-106 0.00 2.52e-45 0.00 ...
#>   ..$ p.sd    : num [1:22] 0.00 6.78e-106 0.00 3.59e-45 0.00 ...
#>  $ importances           :List of 3
#>   ..$ results/synthetic1 :List of 2
#>   .. ..$ intra  : tibble[,12] [11 × 12] (S3: tbl_df/tbl/data.frame)
#>   .. .. ..$ ECM      : num [1:11] NA -0.564 -0.58 -0.244 -0.363 ...
#>   .. .. ..$ ligA     : num [1:11] -0.655 NA -0.67 -0.666 2.264 ...
#>   .. .. ..$ ligB     : num [1:11] 0.296 -0.488 NA -0.476 -0.516 ...
#>   .. .. ..$ ligC     : num [1:11] -0.385 -0.524 -0.431 NA -0.553 ...
#>   .. .. ..$ ligD     : num [1:11] -0.601 2.543 -0.624 -0.629 NA ...
#>   .. .. ..$ prodA    : num [1:11] -0.746 0.838 -0.754 -0.743 0.163 ...
#>   .. .. ..$ prodB    : num [1:11] 0.139 -0.588 2.705 -0.414 -0.571 ...
#>   .. .. ..$ prodC    : num [1:11] -0.593 -0.684 -0.532 2.503 -0.69 ...
#>   .. .. ..$ prodD    : num [1:11] -0.74 0.345 -0.745 -0.735 0.591 ...
#>   .. .. ..$ protE    : num [1:11] 2.454 -0.836 -0.983 0.68 -0.767 ...
#>   .. .. ..$ protF    : num [1:11] -0.858 1.27 -0.837 -0.762 0.329 ...
#>   .. .. ..$ Predictor: chr [1:11] "ECM" "ligA" "ligB" "ligC" ...
#>   .. ..$ para.10: tibble[,12] [11 × 12] (S3: tbl_df/tbl/data.frame)
#>   .. .. ..$ ECM      : num [1:11] NA -0.5785 0.6061 0.0772 -1.4115 ...
#>   .. .. ..$ ligA     : num [1:11] -0.213 NA -1.299 -0.145 1.944 ...
#>   .. .. ..$ ligB     : num [1:11] 1.458 -0.402 NA -0.157 -0.826 ...
#>   .. .. ..$ ligC     : num [1:11] 0.212 -0.26 -0.499 NA -0.54 ...
#>   .. .. ..$ ligD     : num [1:11] 0.322 1.61 -1.387 -0.288 NA ...
#>   .. .. ..$ prodA    : num [1:11] -0.559 0.109 -0.772 -0.75 0.551 ...
#>   .. .. ..$ prodB    : num [1:11] 1.966 -1.165 1.353 -0.345 -0.647 ...
#>   .. .. ..$ prodC    : num [1:11] 0.242 -0.498 -1.056 1.017 -0.475 ...
#>   .. .. ..$ prodD    : num [1:11] -0.679 0.731 -0.683 -0.676 0.215 ...
#>   .. .. ..$ protE    : num [1:11] 0.2747 -0.7324 -1.3513 -0.0159 0.174 ...
#>   .. .. ..$ protF    : num [1:11] -0.676 0.82 -1.076 -0.723 0.737 ...
#>   .. .. ..$ Predictor: chr [1:11] "ECM" "ligA" "ligB" "ligC" ...
#>   ..$ results/synthetic10:List of 2
#>   .. ..$ intra  : tibble[,12] [11 × 12] (S3: tbl_df/tbl/data.frame)
#>   .. .. ..$ ECM      : num [1:11] NA -0.624 -0.61 -0.201 -0.574 ...
#>   .. .. ..$ ligA     : num [1:11] -0.663 NA -0.652 -0.646 2.172 ...
#>   .. .. ..$ ligB     : num [1:11] 0.332 -0.463 NA -0.507 -0.518 ...
#>   .. .. ..$ ligC     : num [1:11] -0.349 -0.475 -0.494 NA -0.518 ...
#>   .. .. ..$ ligD     : num [1:11] -0.666 2.382 -0.647 -0.658 NA ...
#>   .. .. ..$ prodA    : num [1:11] -0.734 0.752 -0.83 -0.815 0.443 ...
#>   .. .. ..$ prodB    : num [1:11] 0.125 -0.581 2.71 -0.359 -0.594 ...
#>   .. .. ..$ prodC    : num [1:11] -0.498 -0.678 -0.489 2.511 -0.701 ...
#>   .. .. ..$ prodD    : num [1:11] -0.737 0.439 -0.788 -0.767 0.865 ...
#>   .. .. ..$ protE    : num [1:11] 2.529 -0.818 -0.962 0.591 -0.583 ...
#>   .. .. ..$ protF    : num [1:11] -0.86 1.315 -0.922 -0.746 0.607 ...
#>   .. .. ..$ Predictor: chr [1:11] "ECM" "ligA" "ligB" "ligC" ...
#>   .. ..$ para.10: tibble[,12] [11 × 12] (S3: tbl_df/tbl/data.frame)
#>   .. .. ..$ ECM      : num [1:11] NA -1.003 0.73 -0.38 -0.737 ...
#>   .. .. ..$ ligA     : num [1:11] -0.351 NA -0.731 -0.409 1.917 ...
#>   .. .. ..$ ligB     : num [1:11] 0.8856 -0.6801 NA -0.0547 -0.6162 ...
#>   .. .. ..$ ligC     : num [1:11] 0.227 -0.458 -0.374 NA -0.458 ...
#>   .. .. ..$ ligD     : num [1:11] -0.275 1.942 -0.473 -0.746 NA ...
#>   .. .. ..$ prodA    : num [1:11] -0.686 0.57 -0.591 -0.738 0.453 ...
#>   .. .. ..$ prodB    : num [1:11] 0.799 -0.883 1.832 -0.257 -0.483 ...
#>   .. .. ..$ prodC    : num [1:11] -0.64 -0.321 -1.23 0.488 -0.53 ...
#>   .. .. ..$ prodD    : num [1:11] -0.3685 0.7708 -0.6248 -0.6569 0.0757 ...
#>   .. .. ..$ protE    : num [1:11] 0.329 -0.531 -0.719 -0.501 -0.493 ...
#>   .. .. ..$ protF    : num [1:11] -0.677 1.384 -0.461 -0.9 0.612 ...
#>   .. .. ..$ Predictor: chr [1:11] "ECM" "ligA" "ligB" "ligC" ...
#>   ..$ results/synthetic2 :List of 2
#>   .. ..$ intra  : tibble[,12] [11 × 12] (S3: tbl_df/tbl/data.frame)
#>   .. .. ..$ ECM      : num [1:11] NA -0.62 -0.538 -0.167 -0.578 ...
#>   .. .. ..$ ligA     : num [1:11] -0.648 NA -0.636 -0.64 2.089 ...
#>   .. .. ..$ ligB     : num [1:11] 0.311 -0.503 NA -0.526 -0.513 ...
#>   .. .. ..$ ligC     : num [1:11] -0.387 -0.541 -0.48 NA -0.552 ...
#>   .. .. ..$ ligD     : num [1:11] -0.659 2.148 -0.694 -0.698 NA ...
#>   .. .. ..$ prodA    : num [1:11] -0.709 0.501 -0.749 -0.724 0.294 ...
#>   .. .. ..$ prodB    : num [1:11] 0.2 -0.547 2.708 -0.418 -0.547 ...
#>   .. .. ..$ prodC    : num [1:11] -0.506 -0.667 -0.545 2.592 -0.635 ...
#>   .. .. ..$ prodD    : num [1:11] -0.726 0.149 -0.775 -0.726 0.438 ...
#>   .. .. ..$ protE    : num [1:11] 2.608 -0.734 -0.849 0.53 -0.629 ...
#>   .. .. ..$ protF    : num [1:11] -0.854 0.833 -0.852 -0.736 0.382 ...
#>   .. .. ..$ Predictor: chr [1:11] "ECM" "ligA" "ligB" "ligC" ...
#>   .. ..$ para.10: tibble[,12] [11 × 12] (S3: tbl_df/tbl/data.frame)
#>   .. .. ..$ ECM      : num [1:11] NA -1.272 0.119 0.517 -1.176 ...
#>   .. .. ..$ ligA     : num [1:11] -0.312 NA -0.316 -0.646 1.975 ...
#>   .. .. ..$ ligB     : num [1:11] 0.269 -0.326 NA -0.212 -0.628 ...
#>   .. .. ..$ ligC     : num [1:11] 0.398 -0.765 -0.277 NA -0.784 ...
#>   .. .. ..$ ligD     : num [1:11] 0.382 1.893 -0.689 -0.597 NA ...
#>   .. .. ..$ prodA    : num [1:11] -0.345 0.227 -0.665 -0.647 0.992 ...
#>   .. .. ..$ prodB    : num [1:11] 0.548 -0.624 1.734 -0.163 -0.755 ...
#>   .. .. ..$ prodC    : num [1:11] -0.224 -0.813 -1.076 0.94 -0.612 ...
#>   .. .. ..$ prodD    : num [1:11] -0.3533 0.9442 -0.6477 -0.6284 0.0517 ...
#>   .. .. ..$ protE    : num [1:11] 0.2708 -0.3901 -0.6065 -0.0898 -0.1631 ...
#>   .. .. ..$ protF    : num [1:11] -0.382 1.029 -0.682 -0.996 0.773 ...
#>   .. .. ..$ Predictor: chr [1:11] "ECM" "ligA" "ligB" "ligC" ...
#>  $ importances.aggregated:List of 2
#>   ..$ intra  :'data.frame':	11 obs. of  12 variables:
#>   .. ..$ ECM      : num [1:11] NA -0.603 -0.576 -0.204 -0.505 ...
#>   .. ..$ ligA     : num [1:11] -0.655 NA -0.652 -0.651 2.175 ...
#>   .. ..$ ligB     : num [1:11] 0.313 -0.484 NA -0.503 -0.516 ...
#>   .. ..$ ligC     : num [1:11] -0.374 -0.513 -0.468 NA -0.541 ...
#>   .. ..$ ligD     : num [1:11] -0.642 2.357 -0.655 -0.662 NA ...
#>   .. ..$ prodA    : num [1:11] -0.73 0.697 -0.778 -0.76 0.3 ...
#>   .. ..$ prodB    : num [1:11] 0.155 -0.572 2.708 -0.397 -0.571 ...
#>   .. ..$ prodC    : num [1:11] -0.532 -0.676 -0.522 2.535 -0.675 ...
#>   .. ..$ prodD    : num [1:11] -0.734 0.311 -0.769 -0.743 0.631 ...
#>   .. ..$ protE    : num [1:11] 2.53 -0.796 -0.931 0.6 -0.66 ...
#>   .. ..$ protF    : num [1:11] -0.857 1.139 -0.87 -0.748 0.439 ...
#>   .. ..$ Predictor: chr [1:11] "ECM" "ligA" "ligB" "ligC" ...
#>   ..$ para.10:'data.frame':	11 obs. of  12 variables:
#>   .. ..$ ECM      : num [1:11] NA -0.9511 0.4848 0.0712 -1.1081 ...
#>   .. ..$ ligA     : num [1:11] -0.292 NA -0.782 -0.4 1.945 ...
#>   .. ..$ ligB     : num [1:11] 0.871 -0.469 NA -0.141 -0.69 ...
#>   .. ..$ ligC     : num [1:11] 0.279 -0.494 -0.383 NA -0.594 ...
#>   .. ..$ ligD     : num [1:11] 0.143 1.815 -0.85 -0.543 NA ...
#>   .. ..$ prodA    : num [1:11] -0.53 0.302 -0.676 -0.712 0.665 ...
#>   .. ..$ prodB    : num [1:11] 1.104 -0.891 1.64 -0.255 -0.628 ...
#>   .. ..$ prodC    : num [1:11] -0.207 -0.544 -1.12 0.815 -0.539 ...
#>   .. ..$ prodD    : num [1:11] -0.467 0.815 -0.652 -0.654 0.114 ...
#>   .. ..$ protE    : num [1:11] 0.292 -0.551 -0.892 -0.202 -0.161 ...
#>   .. ..$ protF    : num [1:11] -0.578 1.078 -0.739 -0.873 0.707 ...
#>   .. ..$ Predictor: chr [1:11] "ECM" "ligA" "ligB" "ligC" ...

# Alternatives
if (FALSE) {

 run_misty(misty.views) %>% collect_results()
 
 collect_results(c("results/sample1", "results/sample2", "results/sample3"))
 
 collect_results("results/sample")
}